requirements.txt
Django==4.1.7       
djangorestframework==3.14.0  
psycopg2==2.9.3        
django-cors-headers==3.10.0  
gunicorn==20.1.0  